[CDS-57897]: Refactored packages
